5.11.0
commit 8c2d7566ba5a36a11d0d1d48c6ff02ea74f446d4
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Wed May 27 11:26:58 2026 +0200

    patch

commit 02438ed9083bb2453d57ca1dd4866e884056028a
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Wed May 27 11:24:38 2026 +0200

    patch

commit 6c4e9c268aea07c0d2c7f3180dea3f8bacdd2c8f
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Wed May 27 11:23:18 2026 +0200

    patch

commit 1bac1fda0f9d985ee7dad6aeefe3b6234711d3bf
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Fri May 22 16:37:43 2026 +0200

    patch

commit 268e49acf98a25ca29da186504c1001259ffabd7
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Fri May 22 10:31:29 2026 +0200

    patch

commit cf820ad83f3f988f7c190fe771d67834b540f259
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Fri May 22 10:26:47 2026 +0200

    patch

commit a8120186ee697e9fadf8c5002d38008b8a80a960
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Fri May 22 10:24:10 2026 +0200

    patch

commit e22f39f44ad0b59a0235d73327740fba0f42a5d6
Merge: d1f06e6 75573ea
Author: Ludo <54670129+lbr38@users.noreply.github.com>
Date:   Thu May 21 18:38:36 2026 +0200

    Merge pull request #375 from Cloud-Kid/improve-empty-states

    Improve empty states

commit d1f06e6e44c25faa84a6d4aa42bb2cc9cad3f733
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Thu May 21 18:37:26 2026 +0200

    patch

commit 358a566ec99e6db474507a05c4d249570a353720
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Tue May 19 17:43:26 2026 +0200

    patch

commit 7e195a189a255169d1db41ad07e71e5d8253b160
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Tue May 19 17:42:18 2026 +0200

    patch

commit 9b9245fdc9255b29ea1a69fbc48119194945413f
Author: Ludovic <54670129+lbr38@users.noreply.github.com>
Date:   Mon May 18 18:48:29 2026 +0200

    5.11.0

commit 1b91d11269c87f697000df11fa2f00d313129277
Merge: de4fe05 ffee6e3
Author: Ludo <54670129+lbr38@users.noreply.github.com>
Date:   Sat May 16 11:09:31 2026 +0200

    Merge pull request #374 from Cloud-Kid/validate-source-repository-templates

    Validate source repository templates in CI

commit 75573ea7a29665db1bedf15e352b03edc4903715
Author: Cloud-Kid <49484832+Cloud-Kid@users.noreply.github.com>
Date:   Mon May 11 11:22:14 2026 +0200

    Improve empty states

commit ffee6e3c675d794854a522819a945adf1e315df1
Author: Cloud-Kid <49484832+Cloud-Kid@users.noreply.github.com>
Date:   Mon May 11 11:22:42 2026 +0200

    Validate source repository templates in CI

+++ www/controllers/Repo/Mirror/Rpm.php
@@ -115,11 +115,11 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
          *  We'll give this modules file a temporary name, to avoid it being included automatically by createrepo_c (it fails every time with modules.yaml file)
          *  It will be renamed and imported by modifyrepo_c later
          */
-        $modulesFileExtension == 'gz' ? $modulesFileTargetName = 'modules-temp.yaml.gz' : null;
-        $modulesFileExtension == 'bz2' ? $modulesFileTargetName = 'modules-temp.yaml.bz2' : null;
-        $modulesFileExtension == 'xz' ? $modulesFileTargetName = 'modules-temp.yaml.xz' : null;
-        $modulesFileExtension == 'zst' ? $modulesFileTargetName = 'modules-temp.yaml.zst' : null;
-        $modulesFileExtension == 'yaml' ? $modulesFileTargetName = 'modules-temp.yaml' : null;
+        $modulesFileExtension == 'gz' ? $modulesFileTargetName = 'modules.yaml.gz' : null;
+        $modulesFileExtension == 'xz' ? $modulesFileTargetName = 'modules.yaml.xz' : null;
+        $modulesFileExtension == 'bz2' ? $modulesFileTargetName = 'modules.yaml.bz2' : null;
+        $modulesFileExtension == 'zst' ? $modulesFileTargetName = 'modules.yaml.zst' : null;
+        $modulesFileExtension == 'yaml' ? $modulesFileTargetName = 'modules.yaml' : null;
 
         /**
          *  Download modules file
@@ -158,9 +158,11 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
             throw new Exception('Error while uncompressing <code>' . $modulesFileTargetName . '</code><br><pre class="codeblock">' . $e->getMessage() . '</pre>');
         }
 
-        // Delete original (uncompressed) file
-        if (!unlink($this->workingDir . '/' . $modulesFileTargetName)) {
-            throw new Exception('Could not delete <code>' . $this->workingDir . '/' . $modulesFileTargetName . '</code>');
+        // Delete original file (only if it was compressed and not already a plain .yaml file)
+        if ($modulesFileExtension != 'yaml') {
+            if (!unlink($this->workingDir . '/' . $modulesFileTargetName)) {
+                throw new Exception('Could not delete <code>' . $this->workingDir . '/' . $modulesFileTargetName . '</code>');
+            }
         }
 
         $this->taskLogSubStepController->completed();
@@ -196,10 +198,10 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
          *  We'll give this updateinfo file a target name according to its file extension
          */
         $updateInfoFileExtension == 'gz' ? $updateInfoFileTargetName = 'updateinfo.xml.gz' : null;
-        $updateInfoFileExtension == 'bz2' ? $updateInfoFileTargetName = 'updateinfo.xml.bz2' : null;
         $updateInfoFileExtension == 'xz' ? $updateInfoFileTargetName = 'updateinfo.xml.xz' : null;
-        $updateInfoFileExtension == 'xml' ? $updateInfoFileTargetName = 'updateinfo.xml' : null;
+        $updateInfoFileExtension == 'bz2' ? $updateInfoFileTargetName = 'updateinfo.xml.bz2' : null;
         $updateInfoFileExtension == 'zst' ? $updateInfoFileTargetName = 'updateinfo.xml.zst' : null;
+        $updateInfoFileExtension == 'xml' ? $updateInfoFileTargetName = 'updateinfo.xml' : null;
 
         /**
          *  Download updateinfo file
@@ -238,9 +240,11 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
             throw new Exception('Error while uncompressing <code>' . $updateInfoFileTargetName . '</code><br><pre class="codeblock">' . $e->getMessage() . '</pre>');
         }
 
-        // Delete original (uncompressed) file
-        if (!unlink($this->workingDir . '/' . $updateInfoFileTargetName)) {
-            throw new Exception('Could not delete <code>' . $this->workingDir . '/' . $updateInfoFileTargetName . '</code>');
+        // Delete original file (only if it was compressed and not already a plain .xml file)
+        if ($updateInfoFileExtension != 'xml') {
+            if (!unlink($this->workingDir . '/' . $updateInfoFileTargetName)) {
+                throw new Exception('Could not delete <code>' . $this->workingDir . '/' . $updateInfoFileTargetName . '</code>');
+            }
         }
 
         $this->taskLogSubStepController->completed();
@@ -539,8 +543,6 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
         if ($error == 0) {
             $this->taskLogSubStepController->completed();
         }
-
-        unset($jsonArray, $data);
     }
 
     /**
@@ -594,10 +596,10 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
              *  If a list of package(s) to include has been provided, check if the package is in the list
              *  If not, skip the package
              */
-            if (!empty($this->packagesToInclude)) {
+            if (!empty($this->advancedParams['packages']['include'])) {
                 $isIn = false;
 
-                foreach ($this->packagesToInclude as $packageToInclude) {
+                foreach ($this->advancedParams['packages']['include'] as $packageToInclude) {
                     if (preg_match('/' . $packageToInclude . '/', $rpmPackageName)) {
                         $isIn = true;
                     }
@@ -616,10 +618,10 @@ class Rpm extends \Controllers\Repo\Mirror\Mirror
              *  If a list of package(s) to exclude has been provided, check if the package is in the list
              *  If so, skip the package
              */
-            if (!empty($this->packagesToExclude)) {
+            if (!empty($this->advancedParams['packages']['exclude'])) {
                 $isIn = false;
 
-                foreach ($this->packagesToExclude as $packageToExclude) {
+                foreach ($this->advancedParams['packages']['exclude'] as $packageToExclude) {
                     if (preg_match('/' . $packageToExclude . '/', $rpmPackageName)) {
                         $isIn = true;
                     }
